Support .gql extension
I would like to support the `.gql` file extension in this addon so that I can differentiate different types of files in my project.  The README already mentions a `.gql` file, but the type must be added to generate the javascript from them.  I've tested this on my local project.